+++ changelog.md
@@ -16,6 +16,17 @@ All notable changes to this project will be documented in this file.
 ### Removed
 -
 -->
+## [1.0.2] - 2024-03-18
+
+### Added
+- Adding setup Makefile
+- Added a requirement for templates only with categories
+- Update the .gitignore file
+
+### Changed
+- All templates require the `name`, `purpose`, and `category` key.
+- Setup script will create a __.py2shell__ file in the home drive.
+
 
 ## [1.0.1] - 2023-04-12
 
@@ -39,5 +50,6 @@ All notable changes to this project will be documented in this file.
 - Initial Commit
 
 <!--  [unreleased]: https://github.com/olivierlacan/keep-a-changelog/compare/v1.0.1...HEAD -->
+[1.0.2]: https://github.com/denezt/automate-shell-scripting/compare/v1.0.1...v1.0.2
 [1.0.1]: https://github.com/denezt/automate-shell-scripting/compare/v1.0.0...v1.0.1
 [1.0.0]: https://github.com/denezt/automate-shell-scripting/releases/tag/v1.0.0